pyproject.toml
[project]
name = "HH_chatbot"
version = "0.1.0"
description = "Chatbot"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.3.0",
    "openai>=1.86.0",
    "pillow>=11.2.1",
    "python-dotenv>=1.1.0",
    "requests>=2.32.4",
]
